feat(config): add usage and storage configurations
* Introduced MAUsageConfig and MAStorageConfig classes for managing energy usage and storage settings.
* Updated LoadConfig to register new configurations.
* Refactored existing configuration handling to accommodate new data types.

+++ src/main/java/io/github/masyumero/mekavaritia/common/tier/transmitter/MATCTier.java
@@ -5,30 +5,30 @@ import mekanism.common.tier.ConductorTier;
 
 public class MATCTier {
     
-    public static long getConduction(ConductorTier tier) {
+    public static double getConduction(ConductorTier tier) {
         return switch (tier) {
-            case BASIC -> LoadConfig.MA_CONFIG.prismaticThermodynamicConductorConduction.get().longValue();
-            case ADVANCED -> LoadConfig.MA_CONFIG.flareThermodynamicConductorConduction.get().longValue();
-            case ELITE -> LoadConfig.MA_CONFIG.neuralThermodynamicConductorConduction.get().longValue();
-            case ULTIMATE -> LoadConfig.MA_CONFIG.eternalThermodynamicConductorConduction.get().longValue();
+            case BASIC -> LoadConfig.MA_CONFIG.prismaticThermodynamicConductorConduction.get();
+            case ADVANCED -> LoadConfig.MA_CONFIG.flareThermodynamicConductorConduction.get();
+            case ELITE -> LoadConfig.MA_CONFIG.neuralThermodynamicConductorConduction.get();
+            case ULTIMATE -> LoadConfig.MA_CONFIG.eternalThermodynamicConductorConduction.get();
         };
     }
 
-    public static long getHeatCapacity(ConductorTier tier) {
+    public static double getHeatCapacity(ConductorTier tier) {
         return switch (tier) {
-            case BASIC -> LoadConfig.MA_CONFIG.prismaticThermodynamicConductornCapacity.get().longValue();
-            case ADVANCED -> LoadConfig.MA_CONFIG.flareThermodynamicConductornCapacity.get().longValue();
-            case ELITE -> LoadConfig.MA_CONFIG.neuralThermodynamicConductornCapacity.get().longValue();
-            case ULTIMATE -> LoadConfig.MA_CONFIG.eternalThermodynamicConductornCapacity.get().longValue();
+            case BASIC -> LoadConfig.MA_CONFIG.prismaticThermodynamicConductornCapacity.get();
+            case ADVANCED -> LoadConfig.MA_CONFIG.flareThermodynamicConductornCapacity.get();
+            case ELITE -> LoadConfig.MA_CONFIG.neuralThermodynamicConductornCapacity.get();
+            case ULTIMATE -> LoadConfig.MA_CONFIG.eternalThermodynamicConductornCapacity.get();
         };
     }
 
-    public static long getConductionInsulation(ConductorTier tier) {
+    public static double getConductionInsulation(ConductorTier tier) {
         return switch (tier) {
-            case BASIC -> LoadConfig.MA_CONFIG.prismaticThermodynamicConductornInsulation.get().longValue();
-            case ADVANCED -> LoadConfig.MA_CONFIG.flareThermodynamicConductornInsulation.get().longValue();
-            case ELITE -> LoadConfig.MA_CONFIG.neuralThermodynamicConductornInsulation.get().longValue();
-            case ULTIMATE -> LoadConfig.MA_CONFIG.eternalThermodynamicConductornInsulation.get().longValue();
+            case BASIC -> LoadConfig.MA_CONFIG.prismaticThermodynamicConductornInsulation.get();
+            case ADVANCED -> LoadConfig.MA_CONFIG.flareThermodynamicConductornInsulation.get();
+            case ELITE -> LoadConfig.MA_CONFIG.neuralThermodynamicConductornInsulation.get();
+            case ULTIMATE -> LoadConfig.MA_CONFIG.eternalThermodynamicConductornInsulation.get();
         };
     }
 }
